+++ src/components/CommonUI/PostModal/PostModal.jsx
@@ -1,11 +1,16 @@
 import { Link } from 'react-router-dom';
+import { useState } from 'react';
 import {
   PostModalArticle,
   GrayLine,
   ModalInnerText,
   CommentModalContainer,
   CommentModalText,
+  ProductModalArticle,
+  ModalBtnWrap,
+  ProductModalText,
 } from './styled';
+import DeleteAlert from '../DeleteAlert/DeleteAlert';
 
 const PostChatModal = ({ leaveChatRoom }) => {
   // 채팅방 나가기
@@ -30,69 +35,39 @@ const CommentModal = ({ content, onClickDelete }) => {
   );
 };
 
-export { PostChatModal, CommentModal };
+const ProductModal = ({ productId, setModal, onModal, productLink }) => {
+  const [onAlert, setOnAlert] = useState(false);
+  const AlertOpen = () => {
+    setOnAlert(!onAlert);
+  };
+  const handleCancel = () => {
+    setOnAlert(false);
+    setModal(false);
+  };
+  const handleCloseModal = () => {
+    setModal(false);
+  };
 
-// export const YourProfileModal = () => {
-//   return (
-//     <PostModalArticle>
-//       <h1 className='ir'>모달창</h1>
-//       <GrayLine />
-//       <Link to='/chatlist'>
-//         <ModalInnerText>설정 및 개인 정보</ModalInnerText>
-//         <ModalInnerText>프로필 설정</ModalInnerText>
-//         <ModalInnerText>로그아웃</ModalInnerText>
-//       </Link>
-//     </PostModalArticle>
-//   );
-// };
+  return (
+    <>
+      <ProductModalArticle onClick={handleCloseModal}>
+        <h2 className='ir'>모달창</h2>
+        <GrayLine />
+        <ModalBtnWrap type='button' onClick={(e) => e.stopPropagation()}>
+          <ProductModalText onClick={AlertOpen}>삭제</ProductModalText>
+          <Link to={`/product/${productId}/modify`}>
+            <ProductModalText>수정</ProductModalText>
+          </Link>
+          <a href={productLink} target='_blank' rel='noopener noreferrer'>
+            <ProductModalText>웹사이트에서 상품 보기</ProductModalText>
+          </a>
+        </ModalBtnWrap>
+      </ProductModalArticle>
+      {onAlert && (
+        <DeleteAlert productId={productId} handleCancel={handleCancel} />
+      )}
+    </>
+  );
+};
 
-// export const MyProfilePostModal = () => {
-//   return (
-//     <PostModalArticle>
-//       <h1 className='ir'>모달창</h1>
-//       <GrayLine />
-//       <Link to='/chatlist'>
-//         <ModalInnerText>삭제</ModalInnerText>
-//         <ModalInnerText>수정</ModalInnerText>
-//       </Link>
-//     </PostModalArticle>
-//   );
-// };
-
-// export const MyProfileProductModal = () => {
-//   return (
-//     <PostModalArticle>
-//       <h1 className='ir'>모달창</h1>
-//       <GrayLine />
-//       <Link to='/chatlist'>
-//         <ModalInnerText>삭제</ModalInnerText>
-//         <ModalInnerText>수정</ModalInnerText>
-//         <ModalInnerText>웹사이트에서 상품 보기</ModalInnerText>
-//       </Link>
-//     </PostModalArticle>
-//   );
-// };
-
-// export const MyProfileProductModal = () => {
-//   return (
-//     <PostModalArticle>
-//       <h1 className='ir'>모달창</h1>
-//       <GrayLine />
-//       <Link to='/chatlist'>
-//         <ModalInnerText>신고하기</ModalInnerText>
-//       </Link>
-//     </PostModalArticle>
-//   );
-// };
-
-// export const MyProfileProductModal = () => {
-//   return (
-//     <PostModalArticle>
-//       <h1 className='ir'>모달창</h1>
-//       <GrayLine />
-//       <Link to='/chatlist'>
-//         <ModalInnerText>삭제</ModalInnerText>
-//       </Link>
-//     </PostModalArticle>
-//   );
-// };
+export { PostChatModal, CommentModal, ProductModal };
