+++ src/components/CommonUI/PostModal/styled.jsx
@@ -61,3 +61,41 @@ export const CommentModalContainer = styled(PostModalArticle)`
 `;
 
 export const CommentModalText = styled(ModalInnerText)``;
+
+export const ProductModalArticle = styled.article`
+  ${({ theme }) => {
+    return css`
+      width: 100%;
+      height: 184px;
+      background-color: #fff;
+      position: fixed;
+      bottom: 0;
+      padding-bottom: 24px;
+      outline: 1px solid ${theme.colors.greyDB};
+      border-radius: ${theme.borderRadius.lv2} ${theme.borderRadius.lv2} 0 0;
+      animation: ${fadeIn} 0.5s;
+      z-index: 10;
+    `;
+  }}
+`;
+
+export const ModalBtnWrap = styled.div`
+  > p {
+    &:nth-child(1) {
+      margin-top: 50px;
+    }
+  }
+`;
+
+export const ProductModalText = styled.p`
+  ${({ theme }) => {
+    return css`
+      font-size: ${theme.fontSizes.md};
+      font-weight: 400;
+      margin-top: 28px;
+      margin-left: 26px;
+      cursor: pointer;
+      color: #000;
+    `;
+  }}
+`;
